Format automatically created resource manager class
